art-o-mat retired cigarette vending machine
In this assignment, you are to create a class specification that simulates an Art-o-mat machine selling artistic works for $0.75 each.  For this assignment, you won't need to write the code that "drives" the machine -- that is provided for you.  Instead, you'll write the class specification so that the provided driver code executes without modification.  That is, you should use the driver code to provide clues about what is needed in your class specification.