finding genes
Write a class Genome that stores a sequence of letters.  Include in your class a constructor that accepts a string to use as the genome, but all lowercase letters are made uppercase, and any extraneous characters are removed. Your class should include a display() method that prints the entire sequence, as well as a genes() method that prints all the genes in the sequence, one per line. If no gene is found in the sequence, the genes() method should print "no gene is found"